+++ breaking-front/src/App.js
@@ -7,12 +7,12 @@ import { ThemeProvider } from 'styled-components';
 import UserInformationProvider from 'providers/UserInformationProvider';
 import { PAGE_PATH } from 'constants/path';
 import Layout from 'components/Layout/Layout';
-import SocialLogin from 'pages/SocialLogin/SocialLogin';
-import KakaoRedirect from 'pages/SocialLogin/Redirect/KakaoRedirect';
-import GoogleRedirect from 'pages/SocialLogin/Redirect/GoogleRedirect';
-import SignUp from 'pages/SignUp/SignUp';
-import Profile from 'pages/Profile/Profile';
-import ProfileEdit from 'pages/ProfileEdit/ProfileEdit';
+import SocialLogin from 'pages/Authentication/SignIn/SocialLogin';
+import KakaoRedirect from 'pages/Authentication/SignIn/Redirect/KakaoRedirect';
+import GoogleRedirect from 'pages/Authentication/SignIn/Redirect/GoogleRedirect';
+import SignUp from 'pages/Authentication/SignUp/SignUp';
+import Profile from 'pages/Profile/Profile/Profile';
+import ProfileEdit from 'pages/Profile/ProfileEdit/ProfileEdit';
 import MainFeed from 'pages/MainFeed/MainFeed';
 import Post from 'pages/Post/Post';
 import PostWrite from 'pages/PostWrite/PostWrite';
