+++ breaking-front/src/pages/Authentication/SignIn/SignIn.js
@@ -1,10 +1,10 @@
 import { GOOGLE_PATH, KAKAO_PATH } from 'constants/path';
 import React from 'react';
 import { useGoogleLogin } from '@react-oauth/google';
-import SocialLoginButton from 'components/SocialLoginButton/SocialLoginButton';
-import * as Style from 'pages/SocialLogin/SocialLogin.styles';
+import SocialLoginButton from 'pages/Authentication/SignIn/components/SocialLoginButton/SocialLoginButton';
+import * as Style from 'pages/Authentication/SignIn/SignIn.styles';
 import Line from 'components/Line/Line';
-import MobileDownloadButton from 'components/MobileDownloadButton/MobileDownloadButton';
+import MobileDownloadButton from 'pages/Authentication/SignIn/components/MobileDownloadButton/MobileDownloadButton';
 
 const SocialLogin = () => {
   const googleLoginClick = useGoogleLogin({
