+++ breaking-front/src/pages/Authentication/SignIn/components/SocialLoginButton/SocialLoginButton.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import * as Style from 'components/SocialLoginButton/SocialLoginButton.styles';
+import * as Style from 'pages/Authentication/SignIn/components/SocialLoginButton/SocialLoginButton.styles';
 
 export default function SocialLoginButton({ social, onClick }) {
   return <Style.LoginButton onClick={onClick} icon={social} />;
